+++ lib/puppet/face/facts/upload.rb
@@ -24,6 +24,9 @@ Puppet::Face.define(:facts, '0.0.1') do
     render_as :json
 
     when_invoked do |options|
+      # Use `agent` sections  settings for certificates, Puppet Server URL,
+      # etc. instead of `user` section settings.
+      Puppet.settings.preferred_run_mode = :agent
       Puppet::Node::Facts.indirection.terminus_class = :facter
       facts = Puppet::Node::Facts.indirection.find(Puppet[:certname])